dirvote: Extract shared functions to common file
No code behavior change.
